Remove file management
The dimensions of the simulation are currently tied to the dimensions
of the graphical view.
I don't want to implement zooming and/or scrolling in this project.
This makes it difficult to save and load a simulation across devices of
different dimensions.
And let's be honest: this is a toy and file saving/loading is not
necessary.